+++ tests/support/stub/ApplicationRest.php
@@ -7,18 +7,18 @@ namespace yii2\extensions\psrbridge\tests\support\stub;
 use Psr\Http\Message\{ResponseInterface, ServerRequestInterface};
 use yii\base\InvalidConfigException;
 use yii\web\IdentityInterface;
-use yii2\extensions\psrbridge\http\{Response, StatelessApplication};
+use yii2\extensions\psrbridge\http\{Application, Response};
 
 /**
  * Stub for a REST application.
  *
  * @template TUserIdentity of IdentityInterface
- * @extends StatelessApplication<TUserIdentity>
+ * @extends Application<TUserIdentity>
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-final class ApplicationRest extends StatelessApplication
+final class ApplicationRest extends Application
 {
     /**
      * Indicates whether `attachPsrRequest()` was invoked.